move menu code to separate file
